Made the CRC a single header file instead of needing a cc file as well.
